review: strip stray BOMs, tighten the probe's positive controls, name the NetBackgroundOp exemption
Findings from /review, all nine fixed:
- the contract claimed no wire byte is ever cast outside the validation
  region, while NetSession's EvBackground case still does. That line belongs
  to a concurrent card, so the exemption is now stated (and its lack of
  Enum.IsDefined coverage called out) instead of the contract being wrong.
- four files had picked up a UTF-8 BOM; no other .cs in the repo has one.
- section 1's stated rationale was wrong (section 2 already subsumes the
  acceptance half); it now says what it actually adds -- the value is
  returned UNCHANGED.
- the EvUnlock and EvCosmeticSwarm positive controls asserted only the bool,
  so a decoder returning true with default(T) passed them; and no row checked
  that a VALID speech cue survives, so a SpeechOrNone hard-wired to Nothing
  passed the whole set. Mutation-tested: that mutant now fails.
- SendLaunch took an int difficulty, forcing an enum->int->byte round trip at
  its only call site.
- the refusal notice named the mission when the tier can be the cause too.
- Players is floored at 0 rather than clamped to MaxPlayers: capping 99 to 4
  made a nonsense row read as a genuinely full game.

+++ web/EvilAliensWeb/Game/EvilAliens/MenuScene.cs
@@ -606,7 +606,7 @@ internal class MenuScene : Scene
 				sender.Remove();
 				return;
 			}
-			EvilAliensWeb.Compat.Net.NetSession.SendLaunch(selectedLevel, (int)Settings.GetInstance().CurrentDifficulty);
+			EvilAliensWeb.Compat.Net.NetSession.SendLaunch(selectedLevel, Settings.GetInstance().CurrentDifficulty);
 		}
 		fadestarted = timer;
 		currentFade = 0f;
